Refactor the formFiels file structure
